% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichodata.R
\docType{data}
\name{dichodata}
\alias{dichodata}
\title{Simulated multiple choice data with 1000 cases, and 20 items}
\description{
Simulated multiple choice data with 1000 cases, and 20 items
}
\keyword{data}
