% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialdata.R
\docType{data}
\name{partialdata}
\alias{partialdata}
\title{Simulated multiple choice data with 1000 cases, and 20 items and partial credits}
\description{
Simulated multiple choice data with 1000 cases, and 20 items and partial credits
}
\keyword{data}
