% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polydata.R
\docType{data}
\name{polydata}
\alias{polydata}
\title{Simulated polytomous items data with 1000 cases, and 20 items}
\description{
Simulated polytomous items data with 1000 cases, and 20 items
}
\keyword{data}
